match second instance of viewport meta
the viewport meta is listed in the essential tags area, and then again in the Meta listing.

In the Meta listing, it has the shrinl-to-fit=no, per necessity for iOS9 devices.

Adding the shrink-to-fit to the first instance of the viewport meta to keep things consistent / not promote a version of the tag that won’t take iOS9 into account.