Remove pagination signals
I saw this is being treated as a curated list. As such, I thought you might want to remove this vestigial signaling code. It's use was proposed by Google, but they [no longer recommend](https://webmasters.googleblog.com/2011/09/pagination-with-relnext-and-relprev.html) it and [haven't used it in years](https://twitter.com/JohnMu/status/1108717999131893765).